# spring-boot/src/main/java/com/example/api/Task.java
package com.example.api;

public class Task {
    private int id;
    private String title;
    private String description;
    private boolean done;

    public Task() {}

    public Task(int id, String title, String description) {
        this.id = id;
        this.title = title;
        this.description = description;
        this.done = false;
    }

    public int getId() { return id; }
    public void setId(int id) { this.id = id; }

    public String getTitle() { return title; }
    public void setTitle(String title) { this.title = title; }

    public String getDescription() { return description; }
    public void setDescription(String description) { this.description = description; }

    public boolean isDone() { return done; }
    public void setDone(boolean done) { this.done = done; }
}